@@ -4,11 +4,11 @@
 <template>
     <div class="folder-view web-view">
         <schemio-header>
-            <div slot="loader">
+            <template v-slot:loader>
                 <div v-if="isLoading" class="loader">
                     <div class="loader-element"></div>
                 </div>
-            </div>
+            </template>
         </schemio-header>
 
         <div class="middle-content">
@@ -65,7 +65,7 @@
                                 </router-link>
                             </td>
                             <td class="time-column">
-                                <span v-if="entry.modifiedTime">{{entry.modifiedTime | formatDateTime }}</span>
+                                <span v-if="entry.modifiedTime">{{$filters.formatDateTime(entry.modifiedTime)}}</span>
                             </td>
                             <td class="operation-column">
                                 <menu-dropdown v-if="entry.name !== '..' && !viewOnly" name="" iconClass="fas fa-ellipsis-v" :options="entry.menuOptions"
@@ -133,6 +133,7 @@ import Modal from '../../components/Modal.vue';
 import CreateNewSchemeModal from '../../components/CreateNewSchemeModal.vue';
 import MenuDropdown from '../../components/MenuDropdown.vue';
 import MoveToFolderModal from '../components/MoveToFolderModal.vue';
+import utils from '../../utils';
 
 const _kindPrefix = (kind) => kind === 'dir' ? 'a': 'b';
 function entriesSorter(a, b) {
@@ -168,20 +169,7 @@ export default {
     },
 
     data() {
-        let path = this.$route.fullPath.trim();
-        if (path === '/') {
-            path = '';
-        }
-        const routePrefix = this.$store.state.routePrefix || '';
-
-        if (path.startsWith(routePrefix)) {
-            path = path.substring(routePrefix.length);
-        }
-
-        if (path.indexOf('/f/') === 0) {
-            path = decodeURI(path.substring(3));
-        }
-
+        const path = this.$route.params.folders ? this.$route.params.folders.join('/') : '';
 
         return {
             path: path,
@@ -318,10 +306,10 @@ export default {
                 this.progressModal.shown = false;
 
                 const routePrefix = this.$store.state.routePrefix || '';
-                if (this.$router.mode === 'history') {
-                    this.$router.push({path: `${routePrefix}/docs/${createdScheme.id}#m=edit`});
+                if (this.$router.options.mode === 'history') {
+                    this.$router.push(`${routePrefix}/docs/${createdScheme.id}#m=edit`);
                 } else {
-                    this.$router.push({path: `${routePrefix}/docs/${createdScheme.id}?m=edit`});
+                    this.$router.push(`${routePrefix}/docs/${createdScheme.id}?m=edit`);
                 }
             })
             .catch(err => {
@@ -443,9 +431,8 @@ export default {
         },
 
         searchSchemes() {
-            this.$router.push({
-                path: `${this.$store.state.routePrefix}/search?q=${encodeURIComponent(this.searchKeyword)}`
-            });
+            const path = `${this.$store.state.routePrefix}/search?q=${encodeURIComponent(this.searchKeyword)}`;
+            this.$router.push(path);
         },
 
         loadNextPage() {
