@@ -4,7 +4,7 @@
 <template>
     <div class="scheme-editor-view" :class="{'diagram-404-view': is404}">
         <schemio-header>
-            <div slot="middle-section">
+            <template v-slot:middle-section>
                 <ul class="header-breadcrumbs">
                     <li v-for="(crumb,  crumbIdx) in breadcrumbs">
                         <router-link v-if="crumb.kind === 'dir'" :to="`${routePrefix}/f/${crumb.path}`">
@@ -19,12 +19,12 @@
                     <img class="icon" :src="`${assetsPath}/images/schemio-logo-white.small.png`" height="20"/>
                     <span>{{scheme.name}}</span>
                 </div>
-            </div>
-            <div slot="loader">
+            </template>
+            <template v-slot:loader>
                 <div v-if="isLoading" class="loader">
                     <div class="loader-element"></div>
                 </div>
-            </div>
+            </template>
         </schemio-header>
 
         <div v-if="is404" class="middle-content">
@@ -179,7 +179,7 @@ export default {
         const schemeId = this.$route.params.schemeId;
 
         return {
-            hasher: createHasher(this.$router ? this.$router.mode : 'history'),
+            hasher: createHasher(this.$router ? this.$router.options.mode : 'history'),
             schemeId: schemeId,
             path: '',
             breadcrumbs: [],
@@ -292,7 +292,7 @@ export default {
                     EditorEventBus.schemeChangeCommitted.$emit(`scheme-${this.appReloadKey}`);
                     this.newSchemePopup.item = null;
                 }
-                const isHistoryMode = this.$router && this.$router.mode === 'history';
+                const isHistoryMode = this.$router && this.$router.options.mode === 'history';
                 const publicLink = isHistoryMode ?  `/docs/${createdScheme.id}#m=edit` : `#/docs/${createdScheme.id}?m=edit`
                 window.open(publicLink, '_blank');
             })
@@ -306,7 +306,7 @@ export default {
         },
 
         loadSchemeFromLink() {
-            const mode = this.$router ? this.$router.mode : 'history';
+            const mode = this.$router ? this.$router.options.mode : 'history';
             const chars = {'.': '+', '_': '/', '-': '='};
 
             const pageParams = createHasher(mode).decodeURLHash(window.location.hash);
