@@ -118,7 +118,7 @@
                     @mouse-double-click="mouseDoubleClick"
                     @svg-size-updated="onSvgSizeUpdated"
                     >
-                    <g slot="scene-transform">
+                    <template v-slot:scene-transform>
                         <EditBox  v-if="schemeContainer.editBox && state !== 'editPath' && state !== 'cropImage' && state !== 'imageBox' && !inPlaceTextEditor.shown"
                             :key="`edit-box-${editorRevision}-${schemeContainer.editBox.id}`"
                             :useFill="state !== 'pickElement' && editBoxUseFill"
@@ -174,9 +174,9 @@
                                 :boundary-box-color="schemeContainer.scheme.style.boundaryBoxColor"
                                 :control-points-color="schemeContainer.scheme.style.controlPointsColor"/>
                         </g>
-                    </g>
+                    </template>
 
-                    <div slot="overlay">
+                    <template v-slot:overlay>
                         <div v-if="state === 'pickElement'" class="editor-top-hint-label">Click any element to pick it</div>
 
                         <FloatingHelperPanel v-if="floatingHelperPanel.shown && floatingHelperPanel.item"
@@ -198,7 +198,7 @@
                             @close="closeStarterProposalModal"
                             @selected="onStarterProposalSelected"
                             />
-                    </div>
+                    </template>
                 </SvgEditor>
 
                 <SvgEditor
@@ -226,12 +226,12 @@
                     @compiler-error="onCompilerError"
                     >
 
-                    <div slot="overlay">
+                    <template v-slot:overlay>
                         <div v-if="mode === 'view' && textSelectionEnabled" class="editor-top-hint-label">
                             You can select any text, but you cannot interact with items
                             <span class="btn btn-primary" @click="textSelectionEnabled = false">Cancel</span>
                         </div>
-                    </div>
+                    </template>
                 </SvgEditor>
 
                 <!-- Item Text Editor -->
@@ -327,8 +327,8 @@
                 <div class="bottom-panel-body">
                     <div class="side-panel-filler-left" :style="{width: `${sidePanelLeftWidth}px`}"></div>
                     <div class="bottom-panel-content">
-                        <span v-if="curveEditing.selectedPoints.length !== 1" class="label">x: {{cursorX | prettifyAxisValue(zoom)}}</span>
-                        <span v-if="curveEditing.selectedPoints.length !== 1" class="label">y: {{cursorY | prettifyAxisValue(zoom)}}</span>
+                        <span v-if="curveEditing.selectedPoints.length !== 1" class="label">x: {{prettifyAxisValue(cursorX, zoom)}}</span>
+                        <span v-if="curveEditing.selectedPoints.length !== 1" class="label">y: {{prettifyAxisValue(cursorY, zoom)}}</span>
                         <div v-if="curveEditing.selectedPoints.length === 1" class="first-selected-point">
                             <span class="label">x: </span>
                             <input type="text" class="textfield"
@@ -432,12 +432,14 @@
                             <span v-if="tab.count">({{tab.count}})</span>
                         </div>
                     </li>
-                    <li v-for="itemTextSlotTab in itemTextSlotsAvailable" v-if="mode === 'edit'">
-                        <span class="tab"
-                            :class="{active: currentTab === itemTextSlotTab.tabName}"
-                            @click="changeTab(itemTextSlotTab.tabName)"
-                            >&#167; {{itemTextSlotTab.slotName}}</span>
-                    </li>
+                    <template v-for="itemTextSlotTab in itemTextSlotsAvailable">
+                        <li v-if="mode === 'edit'">
+                            <span class="tab"
+                                :class="{active: currentTab === itemTextSlotTab.tabName}"
+                                @click="changeTab(itemTextSlotTab.tabName)"
+                                >&#167; {{itemTextSlotTab.slotName}}</span>
+                        </li>
+                    </template>
                     <li v-if="mode !== 'view' && selectedTemplateRef && selectedTemplateRootItem">
                         <span class="tab"
                             :class="{active: currentTab === 'template'}"
@@ -509,22 +511,25 @@
                         </div>
 
                         <div>
-                            <TextSlotProperties v-for="itemTextSlot in itemTextSlotsAvailable" v-if="mode === 'edit' && currentTab === itemTextSlot.tabName"
-                                :key="`text-slot-${itemTextSlot.item.id}-${itemTextSlot.slotName}`"
-                                :editorId="editorId"
-                                :item="itemTextSlot.item"
-                                :slot-name="itemTextSlot.slotName"
-                                @moved-to-slot="onTextSlotMoved(itemTextSlot.item, itemTextSlot.slotName, $event);"
-                                @property-changed="onTextPropertyChanged(itemTextSlot.slotName, $event.name, $event.value)"
+                            <template v-for="itemTextSlot in itemTextSlotsAvailable">
+                                <TextSlotProperties v-if="mode === 'edit' && currentTab === itemTextSlot.tabName"
+                                    :key="`text-slot-${itemTextSlot.item.id}-${itemTextSlot.slotName}`"
+                                    :editorId="editorId"
+                                    :item="itemTextSlot.item"
+                                    :slot-name="itemTextSlot.slotName"
+                                    @moved-to-slot="onTextSlotMoved(itemTextSlot.item, itemTextSlot.slotName, $event);"
+                                    @property-changed="onTextPropertyChanged(itemTextSlot.slotName, $event.name, $event.value)"
+                                    />
+                            </template>
+                            <template v-for="tab in extraTabs">
+                                <component
+                                    v-if="currentTab === `extra:${tab.name}`"
+                                    :key="`tab-${currentTab}`"
+                                    :is="tab.component"
+                                    :mode="mode"
+                                    @tab-event="$emit('custom-tab-event', $event)"
                                 />
-                            <component
-                                v-for="tab in extraTabs"
-                                v-if="currentTab === `extra:${tab.name}`"
-                                :key="`tab-${currentTab}`"
-                                :is="tab.component"
-                                :mode="mode"
-                                @tab-event="$emit('custom-tab-event', $event)"
-                            />
+                            </template>
                         </div>
                     </div>
                 </div>
@@ -1318,7 +1323,7 @@ export default {
             })
             .then(() => {
                 this.isLoading = false;
-                this.schemeContainer = new SchemeContainer(scheme, this.editorId, 'edit', this.$store.state.apiClient, {
+                this.schemeContainer = new SchemeContainer(scheme, this.editorId, this.mode, this.$store.state.apiClient, {
                     onSchemeChangeCommitted: (affinityId) => EditorEventBus.schemeChangeCommitted.$emit(this.editorId, affinityId),
                 });
 
@@ -2828,7 +2833,11 @@ export default {
         },
 
         onContextMenuRequested(mouseX, mouseY, menuOptions) {
-            const svgRect = document.getElementById(`svg-plot-${this.editorId}`).getBoundingClientRect();
+            const svgPlot = document.getElementById(`svg-plot-${this.editorId}`);
+            if (!svgPlot) {
+                return;
+            }
+            const svgRect = svgPlot.getBoundingClientRect();
             this.$emit('context-menu-requested', {
                 x: mouseX + svgRect.left + 5,
                 y: mouseY + svgRect.top + 5,
@@ -3167,7 +3176,11 @@ export default {
             const maxScreen = {x: this._x(max.x), y: this._y(max.y) + bottomMargin};
             const midX = (minScreen.x + maxScreen.x)/2;
 
-            const svgRect = document.getElementById(`svg-plot-${this.editorId}`).getBoundingClientRect();
+            const svgPlot = document.getElementById(`svg-plot-${this.editorId}`)
+            if (!svgPlot) {
+                return;
+            }
+            const svgRect = svgPlot.getBoundingClientRect();
 
             if (midX < 0 || midX > svgRect.width) {
                 this.resetFloatingHelperPanel();
@@ -3668,12 +3681,6 @@ export default {
         y_(y) { return (y - this.schemeContainer.screenTransform.y) / this.schemeContainer.screenTransform.scale },
         z_(v) { return v / this.schemeContainer.screenTransform.scale; },
 
-    },
-
-    filters: {
-        prettifyAxisValue(value, zoom) {
-            return prettifyAxisValue(value, zoom);
-        }
     },
 
     watch: {
