@@ -2,28 +2,14 @@
      License, v. 2.0. If a copy of the MPL was not distributed with this
      file, You can obtain one at https://mozilla.org/MPL/2.0/. -->
 <template lang="html">
-    <div class="rich-text-editor-container">
-        <RichTextMenuBar :editor="editor"/>
-        <div class="editor-frame" :style="{width, height}" :class="{focused}">
-            <div class="editor-content" @click="onEditorClick" :style="{width, height}">
+    <div ref="root" class="rich-text-editor-container" :data-editor-layout="editorLayout">
+        <div class="editor-frame" :style="{width, height: editorContentHeight}" :class="{focused}">
+            <RichTextMenuBar :editor="editor" @link-toggled="onLinkRequested"/>
+            <div class="editor-content" @click="onEditorClick" :style="{width, height: editorContentHeight}">
                 <EditorContent :editor="editor" />
             </div>
-            <span class="text-editor-enlarge" @click="enlarged = true"><i class="fas fa-expand"></i></span>
         </div>
-
-
-        <div class="textarea-enlarged-container" v-if="enlarged && editorLarge">
-            <RichTextMenuBar :editor="editorLarge">
-                <span class="btn btn-secondary action-close" @click="enlarged = false">Close</span>
-            </RichTextMenuBar>
-
-            <div class="editor-frame">
-                <div class="editor-content" @click="onEditorClick">
-                    <EditorContent :editor="editorLarge" />
-                </div>
-            </div>
-        </div>
-
+        <span class="text-editor-enlarge" @click="toggleLayout"><i class="fas fa-expand"></i></span>
 
         <modal v-if="linkModal.shown" title="Add Link" @close="linkModal.shown = false" primary-button="Save" @primary-submit="onLinkModalSubmit">
             <input class="textfield" v-model="linkModal.url" placeholder="https://"  @keyup.enter="onLinkModalSubmit"/>
@@ -32,73 +18,55 @@
 </template>
 
 <script>
-import Link from '@tiptap/extension-link';
 import StarterKit from '@tiptap/starter-kit';
-import { Editor, EditorContent } from '@tiptap/vue-2';
-import Underline from '@tiptap/extension-underline';
+import { Editor, EditorContent } from '@tiptap/vue-3';
 import Modal from '../../Modal.vue';
 import RichTextMenuBar from './RichTextMenuBar.vue';
 import IconTipTapExtension from './IconTipTapExtension';
 
 export default {
     props: {
-        value : {type: String},
-        id    : {type: String},
-        width : {type: String, default: '100%'},
-        height: {type: String, default: '150px'},
+        value      : {type: String},
+        id         : {type: String},
+        width      : {type: String, default: '100%'},
+        height     : {type: String, default: '150px'},
+        linkEnabled: {type: Boolean, default: true},
     },
     components: {Modal, EditorContent, RichTextMenuBar},
 
     beforeMount() {
-        this.editor = new Editor({
+        const editor = new Editor({
             disablePasteRules: true,
             disableInputRules: true,
             extensions: [
                 IconTipTapExtension,
-                StarterKit,
-                Underline,
-                Link.configure({
-                    openOnClick: false,
-                    defaultProtocol: 'https',
+                StarterKit.configure({
+                    link: {
+                        openOnClick: false,
+                        autolink: true,
+                    }
                 }),
             ],
             content: this.value,
             onUpdate: ({ editor }) => {
                 const content = editor.getHTML();
                 this.$emit('changed', content);
-                this.editorLarge.commands.setContent(content, false, {preserveWhitespace: true});
             },
             onFocus: () => {
                 this.focused = true;
             },
             onBlur: () => {
                 this.focused = false;
-            }
+            },
         });
 
-        this.editorLarge = new Editor({
-            disablePasteRules: true,
-            disableInputRules: true,
-            extensions: [
-                IconTipTapExtension,
-                StarterKit,
-                Underline,
-                Link.configure({
-                    openOnClick: false,
-                    defaultProtocol: 'https',
-                }),
-            ],
-            content: this.value,
-            onUpdate: ({ editor }) => {
-                const content = editor.getHTML();
-                this.$emit('changed', content);
-                this.editor.commands.setContent(content, false, {preserveWhitespace: true});
-            }
-        });
+        this.editor = editor;
+
+        document.addEventListener('keydown', this.onKeyPress);
     },
+
     beforeDestroy() {
         this.editor.destroy();
-        this.editorLarge.destroy();
     },
 
     data() {
@@ -106,37 +74,45 @@ export default {
             editor: null,
             focused: false,
             enlarged: false,
-            editorLarge: null,
             enlargedWidth: window.innerWidth - 100,
             enlargedHeight: window.innerHeight - 340,
             linkModal: {
                 shown: false,
                 url: '',
                 onSubmit: null
-            }
+            },
         };
     },
 
     methods: {
+        toggleLayout() {
+            this.enlarged = !this.enlarged;
+        },
+
+        onLinkRequested() {
+            this.toggleLink(this.editor);
+        },
+
         onEditorClick(event) {
             if (event.target.className.indexOf('editor-content') >= 0) {
                 this.editor.view.dom.focus();
             }
         },
 
-        onLargeEditorClick(event) {
-            if (event.target.className.indexOf('editor-content') >= 0) {
-                this.editorLarge.view.dom.focus();
-            }
-        },
-
-
         toggleLink(editor) {
             const previousUrl = this.editor.getAttributes('link').href || '';
             this.linkModal.url = previousUrl;
             this.linkModal.onSubmit = () => {
                 if (this.linkModal.url) {
-                    editor.chain().focus().extendMarkRange('link').setLink({ href: this.linkModal.url }).run();
+                    editor.chain().focus().extendMarkRange('link').run();
+                    if (editor.view.state.selection.empty) {
+                        const link = document.createElement('a');
+                        link.setAttribute('href', this.linkModal.url);
+                        link.innerText = this.linkModal.url;
+                        editor.commands.insertContent(link.outerHTML);
+                    } else {
+                        editor.chain().focus().extendMarkRange('link').setLink({ href: this.linkModal.url }).run();
+                    }
                 } else {
                     editor.chain().focus().extendMarkRange('link').unsetLink().run();
                 }
@@ -147,6 +123,22 @@ export default {
         onLinkModalSubmit() {
             this.linkModal.onSubmit();
             this.linkModal.shown = false;
+        },
+
+        onKeyPress(event) {
+            if (event.key === 'Escape' && this.enlarged) {
+                this.enlarged = false;
+            }
+        },
+    },
+
+    computed: {
+        editorLayout() {
+            return this.enlarged ? 'large' : 'normal';
+        },
+
+        editorContentHeight() {
+            return this.enlarged ? '100%' : this.height;
         }
     },
 
@@ -156,7 +148,6 @@ export default {
         // Otherwise it will not get updated
         id(val) {
             this.editor.commands.setContent(this.value, false, {preserveWhitespace: true});
-            this.editorLarge.commands.setContent(this.value, false, {preserveWhitespace: true});
         }
     }
 }
