@@ -1,21 +1,18 @@
 /* This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-import Vue from 'vue';
-import Vuex from 'vuex';
+import {createStore} from 'vuex';
 import {forEach} from '../collections';
 import utils from '../utils';
 import {createSettingStorageFromLocalStorage} from '../LimitedSettingsStorage';
 import shortid from 'shortid';
 
-Vue.use(Vuex);
-
 const DEFAULT_CONNECTOR_SMOOTHING = 'defaultConnectorSmoothing';
 
 const myStorage = createSettingStorageFromLocalStorage('store', 100);
 
-
-const store = new Vuex.Store({
+const store = createStore({
+    devtools: false,
     state: {
         //rootPath is used in the header for a home link, since if this is being hosted in GitHub Pages then we cannot use '/' as root path
         rootPath: '/',
@@ -74,7 +71,6 @@ const store = new Vuex.Store({
         selectedConnectorPath: null,
         selectedConnector: null,
 
-        multiSelectBox: null,
 
         // used to render snapping lines when user drags item and it is snapped to other items
         snappers: {
@@ -88,7 +84,6 @@ const store = new Vuex.Store({
             connectorItemId: null,
             mx: 0, // value on x viewport axis
             my: 0, // value on y viewport axis
-            primaryShapeId: null,
         },
 
         // used for storing information about images that were dropped on svg editor
@@ -170,10 +165,6 @@ const store = new Vuex.Store({
             state.itemControlPoints.length = 0;
         },
 
-        SET_MULTI_SELECT_BOX(state, box) {
-            state.multiSelectBox = box;
-        },
-
         SET_SELECTED_CONNECTOR_WITH_PATH(state, {item, path}) {
             state.selectedConnectorPath = path;
             state.selectedConnector = item;
@@ -192,12 +183,11 @@ const store = new Vuex.Store({
             state.snappers.vertical = null;
         },
 
-        PROPOSE_CONNECTOR_DESTINATION_ITEMS(state, {connectorItemId, mx, my, primaryShapeId}) {
+        PROPOSE_CONNECTOR_DESTINATION_ITEMS(state, {connectorItemId, mx, my}) {
             state.connectorProposedDestination.connectorItemId = connectorItemId;
             state.connectorProposedDestination.mx = mx;
             state.connectorProposedDestination.my = my;
             state.connectorProposedDestination.shown = true;
-            state.connectorProposedDestination.primaryShapeId = primaryShapeId;
         },
         DISABLE_PROPOSE_CONNECTOR_DESTINATION_ITEMS(state) {
             state.connectorProposedDestination.shown = false;
@@ -422,10 +412,6 @@ const store = new Vuex.Store({
             commit('CLEAR_ITEM_CONTROL_POINTS');
         },
 
-        setMultiSelectBox({commit}, box) {
-            commit('SET_MULTI_SELECT_BOX', box);
-        },
-
         setSelectedConnectorWithPath({commit}, {item, path}) {
             commit('SET_SELECTED_CONNECTOR_WITH_PATH', {item, path});
         },
@@ -449,8 +435,8 @@ const store = new Vuex.Store({
             commit('TOGGLE_ITEM_DETAIL_MARKERS');
         },
 
-        proposeConnectorDestinationItems({commit}, {connectorItemId, mx, my, primaryShapeId}) {
-            commit('PROPOSE_CONNECTOR_DESTINATION_ITEMS', {connectorItemId, mx, my, primaryShapeId});
+        proposeConnectorDestinationItems({commit}, payload) {
+            commit('PROPOSE_CONNECTOR_DESTINATION_ITEMS', payload);
         },
         disableProposeConnectorDestinationItems({commit}) {
             commit('DISABLE_PROPOSE_CONNECTOR_DESTINATION_ITEMS');
@@ -524,8 +510,6 @@ const store = new Vuex.Store({
         apiClient: state => state.apiClient,
 
         itemControlPointsList: state => state.itemControlPoints,
-
-        multiSelectBox: state => state.multiSelectBox,
 
         horizontalSnapper: state => state.snappers.horizontal,
         verticalSnapper: state => state.snappers.vertical,
